+++ mscore/mixer/mixer.cmake
@@ -1,0 +1,23 @@
+
+set(MSCORE_MIXER_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/mixer.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/mixer.h
+    ${CMAKE_CURRENT_LIST_DIR}/mixerdetails.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/mixerdetails.h
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrack.h
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackchannel.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackchannel.h
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackgroup.h
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackitem.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackitem.h
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackpart.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackpart.h
+    )
+
+
+set(MSCORE_MIXER_UI
+    ${CMAKE_CURRENT_LIST_DIR}/mixer.ui
+    ${CMAKE_CURRENT_LIST_DIR}/mixerdetails.ui
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackchannel.ui
+    ${CMAKE_CURRENT_LIST_DIR}/mixertrackpart.ui
+    )
