+++ mscore/mixer/mixerdetails.cpp
