+++ mscore/mixer/mixerdetails.h
