+++ mscore/mixer/mixerdetails.ui
