+++ mscore/mixer/mixertrackchannel.cpp
