+++ mscore/mixer/mixertrackchannel.h
