+++ mscore/mixer/mixertrackitem.cpp
