+++ mscore/parteditbase.cpp
@@ -21,7 +21,7 @@
 #include "parteditbase.h"
 #include "libmscore/score.h"
 #include "libmscore/part.h"
-#include "mixer.h"
+#include "mixer/mixer.h"
 #include "seq.h"
 #include "libmscore/undo.h"
 #include "synthcontrol.h"
