+++ mscore/scoreaccessibility.cpp
@@ -15,7 +15,7 @@
 #include "selectionwindow.h"
 #include "playpanel.h"
 #include "synthcontrol.h"
-#include "mixer.h"
+#include "mixer/mixer.h"
 #include "drumroll.h"
 #include "pianoroll.h"
